# Use tmpfs containers for api/api4 tests, move and speed up CLI tests