--- a/api/api_test.go
+++ b/api/api_test.go
@@ -8,6 +8,8 @@ import (
 	"os"
 	"testing"
 
+	l4g "github.com/alecthomas/log4go"
+
 	"github.com/mattermost/mattermost-server/store/storetest"
 	"github.com/mattermost/mattermost-server/utils"
 )
@@ -18,6 +20,9 @@ func TestMain(m *testing.M) {
 	// In the case where a dev just wants to run a single test, it's faster to just use the default
 	// store.
 	if filter := flag.Lookup("test.run").Value.String(); filter != "" && filter != "." {
+		utils.TranslationsPreInit()
+		utils.LoadConfig("config.json")
+		l4g.Info("-test.run used, not creating temporary containers")
 		os.Exit(m.Run())
 	}