… api/cli_test.go