#### Summary

* The api and api4 tests now use tmpfs MySQL stores if no test filter is used.
* The CLI tests have been moved to cmd/platform so they can be closer to the code that they test (and because getting them to work with an alternative store right now would have been hacky). This requires modification of the "./cmd/platform/*.go" globs everywhere to exclude test files.
* The CLI tests now build an executable before running instead of using "go run". This makes them much faster.
* The api and api4 test helpers now wait on the server to begin accepting connections. This fixes a pre-existing race condition that becomes much more apparent with speed.
* The api and api4 test helpers now ensure that the first user they create is not a system admin. Some tests were failing on a clean database due to that make-the-first-user-a-system-admin logic.

Before (on Jenkins):

```
ok  	github.com/mattermost/mattermost-server/api	401.893s
ok  	github.com/mattermost/mattermost-server/api4	376.092s
```

After:

```
ok  	github.com/mattermost/mattermost-server/api	217.457s
ok  	github.com/mattermost/mattermost-server/api4	257.829s
ok  	github.com/mattermost/mattermost-server/cmd/platform	19.818s
```

This cuts about 5 minutes off the Jenkins build time. I have a lot more to refactor and optimize, but this is a start.

#### Ticket Link
N/A

#### Checklist
N/A